Module Earley.WithPP

A functor providing support for using and Input preprocessor.

Parameters

PP : Input.Preprocessor

Signature

val parse_string : ?⁠filename:string -> 'a grammar -> blank -> string -> 'a

parse_string is the same as Earley.parse_string but it uses the preprocessor defined by PP.

val parse_channel : ?⁠filename:string -> 'a grammar -> blank -> Stdlib.in_channel -> 'a

parse_channel is the same as Earley.parse_channel but it uses the preprocessor defined by PP.

val parse_file : 'a grammar -> blank -> string -> 'a

parse_file is the same as Earley.parse_file but it uses the preprocessor defined by PP.